+++ packages/app-support-buy-sell/src/operations/getOrderDetails/types.ts
@@ -1,0 +1,8 @@
+import { BinanceGetOrderDetailsResponse } from '../../services';
+
+export interface IGetOrderDetailsParams {
+  orderId: string;
+  language?: string;
+}
+
+export type IGetOrderDetailsResult = BinanceGetOrderDetailsResponse['data'];