+++ packages/coin-support-interfaces/src/getAccountAddress.ts
@@ -1,0 +1,5 @@
+import { IAccount } from '@cypherock/db-interfaces';
+
+export interface IGetAccountAddressParams {
+  account: IAccount;
+}