+++ packages/coin-support-starknet/src/operations/receive/index.ts
@@ -21,7 +21,7 @@ import {
 
 import { createApp, deriveAddress } from '../../utils';
 
-const getExternalAddress = async (
+export const getExternalAddress = async (
   params: IGenerateReceiveAddressParams,
 ): Promise<IReceiveAddressInfo> => {
   const { xpubOrAddress, derivationPath } = params.account;