+++ packages/cysync-core/src/pages/MainApp/index.ts
@@ -5,4 +5,5 @@ export * from './Settings';
 export * from './Help';
 export * from './Asset';
 export * from './Account';
+export * from './BuySell';
 export * from './Inheritance';