+++ apps/desktop/extraResources/RELEASE_NOTES.md
@@ -1,3 +1,20 @@
+# v2.0.14
+
+### 🚀 Features
+
+- Added On-ramp: Buy crypto using fiat directly from the cySync app via Binance Connect.
+
+# v2.0.13
+
+### 🚀 Features
+
+- Added support for new EVM tokens
+
+### 🐛 Bugfixes
+
+- Fixed syncing issue with unsupported tokens.
+- Added XRP base reserve API implementation
+
 # v2.0.12
 
 ### 🚀 Features