+++ packages/coin-support-near/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @cypherock/coin-support-near
 
+## 0.0.13
+
+### Patch Changes
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Added get coin address interface
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Add binance connect support
+
+- [#651](https://github.com/Cypherock/cypherock-cysync/pull/651) [`dd105169`](https://github.com/Cypherock/cypherock-cysync/commit/dd1051690015c11fb2da93f885f288e15530e346) Thanks [@TejasvOnly](https://github.com/TejasvOnly)! - Integrate webview
+
+- Updated dependencies [[`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff), [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff), [`dd105169`](https://github.com/Cypherock/cypherock-cysync/commit/dd1051690015c11fb2da93f885f288e15530e346)]:
+  - @cypherock/coin-support-interfaces@0.0.9
+  - @cypherock/coins@0.1.1
+  - @cypherock/cysync-utils@0.0.9
+  - @cypherock/coin-support-utils@0.0.13
+
+## 0.0.13-pre.0
+
+### Patch Changes
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Added get coin address interface
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Add binance connect support
+
+- Updated dependencies [[`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff), [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff)]:
+  - @cypherock/coin-support-interfaces@0.0.9-pre.0
+  - @cypherock/coins@0.1.1-pre.0
+  - @cypherock/cysync-utils@0.0.9-pre.0
+  - @cypherock/coin-support-utils@0.0.13-pre.0
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`1820213c`](https://github.com/Cypherock/cypherock-cysync/commit/1820213c8c9f6d173b3e245758704d10d62ccf1a), [`b9857909`](https://github.com/Cypherock/cypherock-cysync/commit/b985790932df37f970b2568f9bea9115a44aa742)]:
+  - @cypherock/coins@0.1.0
+  - @cypherock/db-interfaces@0.1.0
+  - @cypherock/coin-support-utils@0.0.12
+  - @cypherock/coin-support-interfaces@0.0.8
+  - @cypherock/cysync-interfaces@0.0.7
+  - @cypherock/cysync-utils@0.0.8
+
+## 0.0.12-pre.0
+
+### Patch Changes
+
+- Updated dependencies [[`1820213c`](https://github.com/Cypherock/cypherock-cysync/commit/1820213c8c9f6d173b3e245758704d10d62ccf1a), [`b9857909`](https://github.com/Cypherock/cypherock-cysync/commit/b985790932df37f970b2568f9bea9115a44aa742)]:
+  - @cypherock/coins@0.1.0-pre.0
+  - @cypherock/db-interfaces@0.1.0-pre.0
+  - @cypherock/coin-support-utils@0.0.12-pre.0
+  - @cypherock/coin-support-interfaces@0.0.8-pre.0
+  - @cypherock/cysync-interfaces@0.0.7-pre.0
+  - @cypherock/cysync-utils@0.0.8-pre.0
+
 ## 0.0.11
 
 ### Patch Changes