+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cypherock/cysync-ui
 
+## 0.0.9
+
+### Patch Changes
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Added get coin address interface
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Add binance connect support
+
+- [#651](https://github.com/Cypherock/cypherock-cysync/pull/651) [`dd105169`](https://github.com/Cypherock/cypherock-cysync/commit/dd1051690015c11fb2da93f885f288e15530e346) Thanks [@TejasvOnly](https://github.com/TejasvOnly)! - Integrate webview
+
+## 0.0.9-pre.0
+
+### Patch Changes
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Added get coin address interface
+
+- [#541](https://github.com/Cypherock/cypherock-cysync/pull/541) [`1b1330cf`](https://github.com/Cypherock/cypherock-cysync/commit/1b1330cf125146484d927f4ac8f5eeab466346ff) Thanks [@irshadjsr21](https://github.com/irshadjsr21)! - Add binance connect support
+
 ## 0.0.8
 
 ### Patch Changes